Add files via upload
Volatility measures the extent of variation in the returns of a security or market index and is often used to assess risk. Higher volatility implies greater risk, while lower volatility suggests less risk. GARCH models are extensively used in finance because they effectively capture the changing nature of volatility over time. In this tutorial, we will develop an application using real financial data.