Add files via upload
A Smart Beta Portfolio aims to improve upon traditional market-capitalization-weighted indexing by using alternative weighting strategies. Instead of simply following the market, it relies on certain factors or fundamentals, such as dividends or trading volumes, to determine the weights of the stocks in the portfolio. This approach seeks to provide better risk-adjusted returns. Using the cvxpy library, the we perform a quadratic optimization to minimize the portfolio variance while keeping the weights close to the initial benchmark weights. Logically, this involves solving an optimization problem with constraints that the weights must be non-negative and sum up to 1. Then, it periodically rebalances the portfolio to optimize returns while managing risks.